+++ src/Converter/Number/BigNumberConverter.php
@@ -19,7 +19,7 @@ use Ramsey\Uuid\Math\BrickMathCalculator;
 
 /**
  * Previously used to integrate moontoast/math as a bignum arithmetic library,
- * BigNumberConverter is deprecated in favor of ArbitraryPrecisionNumberConverter
+ * BigNumberConverter is deprecated in favor of GenericNumberConverter
  *
  * @deprecated Transition to {@see GenericNumberConverter}.
  */