+++ src/Converter/Time/PhpTimeConverter.php
@@ -16,7 +16,10 @@ namespace Ramsey\Uuid\Converter\Time;
 
 use Ramsey\Uuid\Converter\TimeConverterInterface;
 use Ramsey\Uuid\Math\BrickMathCalculator;
+use Ramsey\Uuid\Math\CalculatorInterface;
+use Ramsey\Uuid\Type\Hexadecimal;
 use Ramsey\Uuid\Type\IntegerValue;
+use Ramsey\Uuid\Type\Time;
 
 /**
  * PhpTimeConverter uses built-in PHP functions and standard math operations
@@ -26,17 +29,35 @@ use Ramsey\Uuid\Type\IntegerValue;
 class PhpTimeConverter implements TimeConverterInterface
 {
     /**
+     * @var CalculatorInterface
+     */
+    private $calculator;
+
+    /**
      * @var TimeConverterInterface
      */
     private $fallbackConverter;
 
-    public function __construct(?TimeConverterInterface $fallbackConverter = null)
-    {
-        if ($fallbackConverter === null) {
-            $fallbackConverter = new GenericTimeConverter(new BrickMathCalculator());
+    /**
+     * @var int
+     */
+    private $phpPrecision;
+
+    public function __construct(
+        ?CalculatorInterface $calculator = null,
+        ?TimeConverterInterface $fallbackConverter = null
+    ) {
+        if ($calculator === null) {
+            $calculator = new BrickMathCalculator();
         }
 
+        if ($fallbackConverter === null) {
+            $fallbackConverter = new GenericTimeConverter($calculator);
+        }
+
+        $this->calculator = $calculator;
         $this->fallbackConverter = $fallbackConverter;
+        $this->phpPrecision = (int) ini_get('precision');
     }
 
     /**
@@ -79,18 +100,57 @@ class PhpTimeConverter implements TimeConverterInterface
      * @inheritDoc
      * @psalm-pure
      */
-    public function convertTime(string $timestamp): string
+    public function convertTime(Hexadecimal $uuidTimestamp): Time
     {
-        $timestamp = new IntegerValue($timestamp);
+        $timestamp = $this->calculator->toIntegerValue($uuidTimestamp);
 
-        $unixTime = ((int) $timestamp->toString() - 0x01b21dd213814000) / 10000000;
+        $splitTime = $this->splitTime(
+            ((int) $timestamp->toString() - 0x01b21dd213814000) / 10000000
+        );
 
-        if (!is_int($unixTime)) {
-            return $this->fallbackConverter->convertTime(
-                $timestamp->toString()
-            );
+        if (count($splitTime) === 0) {
+            return $this->fallbackConverter->convertTime($uuidTimestamp);
         }
 
-        return (string) $unixTime;
+        return new Time($splitTime['sec'], $splitTime['usec']);
+    }
+
+    /**
+     * @param int|float $time The time to split into seconds and microseconds
+     *
+     * @return string[]
+     *
+     * @psalm-pure
+     */
+    private function splitTime($time): array
+    {
+        $split = explode('.', (string) $time, 2);
+
+        // If the $time value is a float but $split only has 1 element, then the
+        // float math was rounded up to the next second, so we want to return
+        // an empty array to allow use of the fallback converter.
+        if (is_float($time) && count($split) === 1) {
+            return [];
+        }
+
+        if (count($split) === 1) {
+            return [
+                'sec' => $split[0],
+                'usec' => '0',
+            ];
+        }
+
+        // If the microseconds are less than six characters AND the length of
+        // the number is greater than or equal the PHP precision , then it's
+        // possible that we lost some precision for the microseconds. Return an
+        // empty array, so that we can choose to use the fallback converter.
+        if (strlen($split[1]) < 6 && strlen((string) $time) >= $this->phpPrecision) {
+            return [];
+        }
+
+        return [
+            'sec' => $split[0],
+            'usec' => str_pad($split[1], 6, '0', STR_PAD_RIGHT),
+        ];
     }
 }